added role

+++ src/posts/posts.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, ForbiddenException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, In } from 'typeorm';
 import { PostEntity } from './entities/post.entity';
 import { CategoryEntity } from '../categories/entities/category.entity';
 import { UserEntity } from '../users/entities/user.entity';
+import { UserRole } from '../users/enums/user-role.enum';
 import { CreatePostInput } from './dto/create-post.input';
 import { UpdatePostInput } from './dto/update-post.input';
 
@@ -18,14 +19,14 @@ export class PostsService {
     private usersRepository: Repository<UserEntity>,
   ) {}
 
-  async create(createPostInput: CreatePostInput): Promise<PostEntity> {
-    const { authorId, categoryIds, ...postData } = createPostInput;
+  async create(createPostInput: CreatePostInput, userId: number): Promise<PostEntity> {
+    const { categoryIds, ...postData } = createPostInput;
 
     const author = await this.usersRepository.findOne({
-      where: { id: authorId },
+      where: { id: userId },
     });
     if (!author) {
-      throw new NotFoundException(`User with ID ${authorId} not found`);
+      throw new NotFoundException(`User with ID ${userId} not found`);
     }
 
     let categories: CategoryEntity[] = [];
@@ -63,21 +64,15 @@ export class PostsService {
     return post;
   }
 
-  async update(updatePostInput: UpdatePostInput): Promise<PostEntity> {
-    const { id, categoryIds, authorId, ...updateData } = updatePostInput;
+  async update(updatePostInput: UpdatePostInput, user: UserEntity): Promise<PostEntity> {
+    const { id, categoryIds, ...updateData } = updatePostInput;
     const post = await this.findOne(id);
 
-    Object.assign(post, updateData);
-
-    if (authorId) {
-      const author = await this.usersRepository.findOne({
-        where: { id: authorId },
-      });
-      if (!author) {
-        throw new NotFoundException(`User with ID ${authorId} not found`);
-      }
-      post.author = author;
+    if (user.role !== UserRole.ADMIN && post.author.id !== user.id) {
+      throw new ForbiddenException('You can only update your own posts');
     }
+
+    Object.assign(post, updateData);
 
     if (categoryIds) {
       post.categories = await this.categoriesRepository.findBy({
@@ -88,8 +83,13 @@ export class PostsService {
     return await this.postsRepository.save({ ...post, id: +post.id });
   }
 
-  async remove(id: number): Promise<PostEntity> {
+  async remove(id: number, user: UserEntity): Promise<PostEntity> {
     const post = await this.findOne(id);
+
+    if (user.role !== UserRole.ADMIN && post.author.id !== user.id) {
+      throw new ForbiddenException('You can only delete your own posts');
+    }
+
     await this.postsRepository.remove(post);
     return post;
   }
